[后端 bug] order 中 quantity 的类型转换

+++ backend/app/order/biz/service/list_oder.go
@@ -28,7 +28,7 @@ func (s *ListOderService) Run(req *order.ListOrderReq) (resp *order.ListOrderRes
 				Cost: vitem.Cost,
 				Item: &cart.CartItem{
 					ProductId: vitem.ProductId,
-					Quantity:  vitem.Quantity,
+					Quantity:  uint32(vitem.Quantity),
 				},
 			})
 		}
