+++ backend/app/order/biz/service/place_order.go
@@ -54,7 +54,7 @@ func (s *PlaceOrderService) Run(req *order.PlaceOrderReq) (resp *order.PlaceOrde
 			items = append(items, model.OrderItem{
 				OrderIdRefer: orderId.String(),
 				ProductId:    v.Item.ProductId,
-				Quantity:     v.Item.Quantity,
+				Quantity:     int32(v.Item.Quantity),
 				Cost:         v.Cost,
 			})
 		}
